requirements.txt
altgraph==0.17.4
amqp==5.1.1
ansi2html==1.8.0
beautifulsoup4==4.12.2
billiard==4.1.0
bottle==0.13.1
bottle-websocket==0.2.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.7
diskcache==5.6.1
Eel==0.17.0
et-xmlfile==1.1.0
Flask==2.2.5
fonttools==4.41.1
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.4
imageio==2.36.0
imbalanced-learn==0.12.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.4
kombu==5.3.1
lazy_loader==0.4
Levenshtein==0.21.1
macholib==1.16.3
MarkupSafe==2.1.3
matplotlib==3.9.2
multiprocess==0.70.15
nest-asyncio==1.5.7
networkx==3.4.2
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.2.3
Pillow==10.0.0
plotly==5.15.0
prompt-toolkit==3.0.39
psutil==5.9.5
pydash==7.0.6
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.21.1
pytz==2023.3
rapidfuzz==3.1.2
requests==2.31.0
retrying==1.3.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.4.1
tenacity==8.2.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
whichcraft==0.6.1
XlsxWriter==3.1.2
zope.event==5.0
zope.interface==7.0.3
